requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
black==24.3.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
consumers==0.6.1
cryptography==42.0.5
daphne==4.1.0
Django==5.0.4
django-jazzmin==2.6.1
djangorestframework==3.15.1
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
zope.interface==6.2